backend/requirements.txt
Flask>=2.3.0,<3.1.0
flask-cors>=4.0.0
qrcode>=7.4.2
Pillow>=10.0.0
Werkzeug>=2.3.0,<3.1.0
pywhatkit>=5.4
pyautogui>=0.9.54
flask
gunicorn
flask-cors
sqlalchemy
psycopg2-binary
dotenv
